Better CmdSpec computing
Makes the config a private member as it is now accessed only from the
summon package.

We normalize earlier the invokers and handles. Because handles are
unique, we make a global map of them, derived from the config. We now
use these normalized structs throughout. This removes CmdSpec computing
duplication.

This also allows failing fast if duplicated handle are detected.