Driver is the lone responsible for setting args
Refactor command build to delegate more parsing to the cobra library.
This simplifies out the parallel command structure between CmdSpec and
cobra.Command tree. Argument parsing is thus more robust and test
arguments construction is unified.

We can now mostly ignore the presence or absence of the run command as
it is encoded in the cobra.Command tree. (The firsUnknownArgPos special
and annoying case is removed).

Normalized config (CmdSpec and Handles) belong to the summon package
and have been moved there.

Added a file to store flag operations (flag.go in summon package)