Allows disabling summonables as commands in config
The effect of this is that they do not show up in help, but are shown in
completions.

Fix bug where summonable commands were added to the wrong sub-command
(list command). This was missed because there were no tests with run
and without runl subcommands. This is addressed.

Ensure that completion still works if summonables are not commands.

Hides the help subcommand because we have a --help flag.

Fixes root_tests that were written in a way that they were gathering
state in the root creation helper (was called at test table creation
time).