enhances debug of called commands
to enable this, a better global flag parsing was needed to fix a
regression where global flags were not parsed on sub commands.

for instance, debug was not taken into account for run commands.

This is also a preamble to fix proxying know flags like `-d`) as unkown flags
to sub-commands